server/requirements.txt
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==1.4.31
stripe==11.1.0
tomli==2.2.1
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==2.0.3
